Print to stdout
A couple of changes and you can see it working now. I made a lot of mistakes - would be nice to read it off a spreadsheet. Until then, would be good to somehow keep the name of the last tutor given somewhere, so you can repeat it by pressing Enter, say, if two consecutive students have the same tutor.